DT-538: Simplify how equipments are handled when confirming the booking
For historical reasons, we did a complex equipment assignment setup
due to our usage of the "information model as SQL".  This is no longer
required and we can just directly receive and store the confirmed
equipments.  Additionally, all of this very complicated assignment
code was not tested.

This clean up goes the extra mile to get the confirmed equipment
mapped and align a few attributes with the current Swagger API.  There
are still differences between the RI and Swagger (we are not going the
extra 5 miles).

+++ edocumentation-transfer-obj/src/main/java/org/dcsa/edocumentation/transferobjects/RequestedEquipmentTO.java
@@ -26,7 +26,7 @@ public record RequestedEquipmentTO(
   WeightUnit tareWeightUnit,
 
   @NotNull
-  Float units,
+  Integer units,
 
   List<@NotBlank @ISO6346EquipmentReference String> equipmentReferences,
